+++ exonum-java-binding/core/rust/src/storage/mod.rs
@@ -12,9 +12,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+pub use self::access::Java_com_exonum_binding_core_storage_database_Accesses_nativeFree;
+pub(crate) use self::access::{into_generic_raw_access, EjbAccessExt};
 pub use self::blockchain::*;
-pub use self::db::Java_com_exonum_binding_core_storage_database_Accesses_nativeFree;
-pub(crate) use self::db::{into_generic_raw_access, EjbAccessExt};
 pub use self::key_set_index::*;
 pub use self::list_index::*;
 pub use self::map_index::*;
@@ -27,8 +27,8 @@ pub use self::raw_proof_map_index::*;
 pub use self::temporarydb::*;
 pub use self::value_set_index::*;
 
+mod access;
 mod blockchain;
-mod db;
 mod entry;
 mod fork;
 mod key_set_index;
@@ -42,3 +42,6 @@ mod proof_map_index_next;
 mod raw_proof_map_index;
 mod temporarydb;
 mod value_set_index;
+
+pub(crate) type Key = Vec<u8>;
+pub(crate) type Value = Vec<u8>;
