+++ exonum-java-binding/core/rust/src/storage/proof_map_index_next.rs
@@ -28,7 +28,7 @@ use jni::{
 use JniResult;
 
 use handle::{self, Handle};
-use storage::{db::Value, PairIter};
+use storage::{PairIter, Value};
 use utils;
 
 type Key = Vec<u8>;
